docready_api/requirements.txt
Django==1.5.1
argparse==1.2.1
distribute==0.6.34
django-cors-headers==0.10
djangorestframework==2.3.1
wsgiref==0.1.2
